SPARK Silver Proof for the Rover.Sonar Unit
This commit demonstrates successful proof at SPARK Silver for the
Rover.Sonar unit.

We introduce a `Rover.HW` package to provide a SPARK-compatible
interface to the rover hardware. This package should be extended to
enable additional proofs of other Rover units.

Note the hard-coding of the proof level in the `.gpr` file; this should
be removed during development to allow use specified proof levels.