Prove rover controllers cannot command a crash
This commit introduces a formalization of the safety requirement:

> The Mars Rover shall not proceed straight forward when the distance to
> an obstacle is less than the Safety Threshold.

via a new predicate named `Cannot_Crash`:

```ada
   function Cannot_Crash return Boolean is
     (if Rover_HAL.Get_Sonar_Distance < Safety_Distance and then
         Rover_HAL.Get_Turn = Rover_HAL.Straight
      then
         Rover_HAL.Get_Power (Rover_HAL.Left)  <= 0 and then
         Rover_HAL.Get_Power (Rover_HAL.Right) <= 0);
```

This predicate is checked in all top-level subprogram postconditions and
in all relevant loops within the autonomous and remote-controlled
controllers. By provind these properties, we show that at no time do the
controllers for the rover command unsafe behavior.

The predicate, which is Ghost, relies on Ghost functions introduced into
the Rover_HAL that model retrieving the last-command sent to turn or set
motor power and the last distance obtained from the sonar. In this way,
we are able to reason about what the controllers are doing without
modifying the HAL implementation in any way or introducing explicit
global Ghost state.

All checks prove at level 1 with GNATprove FSF