Fix the proof in Remote Control
The model for `Get_Sonar_Distance` was incomplete - it did not include
necessary abstract state (as was added to `Get_Turn` and `Get_Power`).
As a result, the model was overconstrained and we were able to prove the
safety property when we should not have been able to do so.

This commit fixes the issue and introduces a fix to the remote control
code that in fact makes it safer than it was before.